pyproject.toml
[project]
name = "camcot-experiments"
version = "0.1.0"
description = "CAM-CoT experiments"
requires-python = ">=3.10"

dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "wandb>=0.16.0",
  "torch>=2.2.0",
  "transformers>=4.38.0",
  "datasets>=2.16.0",
  "optuna>=3.5.0",
  "numpy>=1.26.0",
  "pandas>=2.1.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
